requirements.txt
canto-explain @ file:///Users/david/.Trash/canto_explain/dist/canto_explain-0.1.0-py3-none-any.whl#sha256=151809c4ae082db0c2b20840f5ad83d37aaca650b377eee794a0760a117e9d1e
Cantonese==1.0.10
cantonese-romanisation==1.0.5
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
packaging==25.0
pillow==12.0.0
pycantonese==3.4.0
Pygments==2.19.2
pylangacq==0.16.2
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
QtPy==2.4.3
requests==2.32.5
setuptools==80.9.0
shiboken6==6.10.0
six==1.17.0
tabulate==0.9.0
urllib3==2.5.0
wcwidth==0.2.14
wordseg==0.0.2
zhconv==1.4.3
